+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/service/rds/MemberServiceAsRds.java
@@ -9,7 +9,7 @@ import com.sparta.moim.gathering.gathering.application.exception.RoleNotAllowedG
 import com.sparta.moim.gathering.gathering.application.service.MemberService;
 import com.sparta.moim.gathering.gathering.domain.repository.GatheringValidationRepository;
 import com.sparta.moim.gathering.gathering.domain.repository.MemberRepository;
-import com.sparta.moim.gathering.sherd.enums.MemberType;
+import com.sparta.moim.gathering.shared.enums.MemberType;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
