+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/infrastructure/repository/JpaMemberRepository.java
@@ -2,7 +2,7 @@ package com.sparta.moim.gathering.gathering.infrastructure.repository;
 
 import com.sparta.moim.gathering.gathering.domain.entity.Member;
 import com.sparta.moim.gathering.gathering.domain.repository.MemberRepository;
-import com.sparta.moim.gathering.sherd.enums.MemberType;
+import com.sparta.moim.gathering.shared.enums.MemberType;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
