+++ gathering-service/src/main/java/com/sparta/moim/gathering/shared/enums/MemberType.java
@@ -1,0 +1,5 @@
+package com.sparta.moim.gathering.shared.enums;
+
+public enum MemberType {
+  ADMIN, GENERAL
+}
