--- gathering-service/src/main/java/com/sparta/moim/gathering/sherd/enums/MemberType.java
@@ -1,5 +1,0 @@
-package com.sparta.moim.gathering.sherd.enums;
-
-public enum MemberType {
-  ADMIN, GENERAL
-}
